restructure metmetpy module
