+++ cdm_reader_mapper/__init__.py
@@ -19,10 +19,12 @@ from .duplicates.duplicates import DupDetect  # noqa
 from .duplicates.duplicates import duplicate_check  # noqa
 from .mdf_reader.reader import read_data, read_mdf  # noqa
 from .mdf_reader.writer import write_data  # noqa
-from .metmetpy.datetime.correct import correct as correct_datetime  # noqa
-from .metmetpy.datetime.validate import validate as validate_datetime  # noqa
-from .metmetpy.platform_type.correct import correct as correct_pt  # noqa
-from .metmetpy.station_id.validate import validate as validate_id  # noqa
+from .metmetpy import (  # noqa
+    correct_datetime,
+    correct_pt,
+    validate_datetime,
+    validate_id,
+)
 
 __author__ = """Ludwig Lierhammer"""
 __email__ = "ludwiglierhammer@dwd.de"
