+++ cdm_reader_mapper/metmetpy/__init__.py
@@ -3,7 +3,5 @@
 from __future__ import annotations
 
 from . import properties  # noqa
-from .datetime import correct as correct_datetime  # noqa
-from .datetime import validate as validate_datetime  # noqa
-from .platform_type import correct as correct_pt  # noqa
-from .station_id import validate as validate_id  # noqa
+from .correct import correct_datetime, correct_pt  # noqa
+from .validate import validate_datetime, validate_id  # noqa
