+++ cdm_reader_mapper/metmetpy/validate.py
@@ -1,7 +1,31 @@
 """
-Validate ID field in a pandas DataFrame.
+metmetpy validation package.
 
 Created on Tue Jun 25 09:00:19 2019
+
+Validates the datetime fields of a data model:
+    -1. extracts or creates the datetime field of a data model as defined
+    in submodule model_datetimes.
+    -2. validates to False where NaT: no datetime or conversion to datetime failure
+
+Validation is data model specific.
+
+Output is a boolean series.
+
+Does not account for input dataframes/series stored in TextParsers: as opposed
+to correction modules, the output is only a boolean series which is external
+to the input data ....
+
+If the datetime conversion (or extraction) for a given data model is not
+available in submodule model_datetimes, the module
+will return with no output (will break full processing downstream of its
+invocation) logging an error.
+
+Reference names of different metadata fields used in the metmetpy modules
+and its location column|(section,column) in a data model are
+registered in ../properties.py in metadata_datamodels.
+
+NaN, NaT: will validate to False.
 
 Validates ID field in a pandas dataframe against a list of regex patterns.
 Output is a boolean series.
@@ -32,17 +56,15 @@ will warn and validate all to True, with NaN to False
 
 from __future__ import annotations
 
-import re
-
 import pandas as pd
 
 from cdm_reader_mapper.common import logging_hdlr, pandas_TextParser_hdlr
-from cdm_reader_mapper.common.json_dict import collect_json_files, combine_dicts
 
-from .. import properties
+from .datetime import model_datetimes
+
+from . import properties
 
 _base = f"{properties._base}.station_id"
-
 
 def _get_id_col(data, imodel, logger):
     id_col = properties.metadata_datamodels["id"].get(imodel)
@@ -83,7 +105,7 @@ def _get_patterns(dck_id_model, blank, dck, data_model_files, logger):
     return patterns
 
 
-def validate(data, imodel, blank=False, log_level="INFO"):
+def validate_id(data, imodel, blank=False, log_level="INFO"):
     """DOCUMENTATION."""
     logger = logging_hdlr.init_logger(__name__, level=log_level)
     if not isinstance(data, pd.DataFrame) and not isinstance(data, pd.Series):
@@ -127,3 +149,33 @@ def validate(data, imodel, blank=False, log_level="INFO"):
     combined_compiled = re.compile("|".join(patterns))
 
     return idSeries.str.match(combined_compiled, na=na_values)
+
+def validate_datetime(data, imodel, log_level="INFO"):
+    """DOCUMENTATiON."""
+    # dck input only to be consistent with other validators in the metmetpy module
+    logger = logging_hdlr.init_logger(__name__, level=log_level)
+    model = imodel.split("_")[0]
+
+    if isinstance(data, pd.io.parsers.TextFileReader):
+        data = pandas_TextParser_hdlr.make_copy(data).read()
+    elif not isinstance(data, pd.DataFrame) and not isinstance(data, pd.Series):
+        logger.error(
+            f"Input data must be a pd.DataFrame or pd.Series.\
+                     Input data type is {type(data)}"
+        )
+        return
+
+    data_model_datetime = model_datetimes.to_datetime(data, model)
+
+    if not isinstance(data_model_datetime, pd.Series):
+        logger.error(
+            f'Data model "{model}" datetime conversor not defined in model_datetimes module"'
+        )
+        return
+    elif len(data_model_datetime) == 0:
+        data_columns = list(data.columns)
+        logger.info(
+            f"No columns found for datetime conversion. Selected columns are {data_columns}"
+        )
+        return
+    return data_model_datetime.notna()
